% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{cent.1}
\alias{cent.1}
\title{Centering coefficients for triaxial nnetinputs}
\format{
A named numeric vector
}
\usage{
cent.1
}
\description{
Centering coefficients for triaxial nnetinputs
}
\keyword{datasets}
