% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{cent}
\alias{cent}
\title{Centering coefficients for uniaxial nnetinputs}
\format{
A named numeric vector
}
\usage{
cent
}
\description{
Centering coefficients for uniaxial nnetinputs
}
\keyword{datasets}
