% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{reg.nn}
\alias{reg.nn}
\title{Uniaxial neural network for use in the original uniaxial Sojourn method}
\format{
From print(reg.nn):
a 6-25-1 network with 207 weights inputs: X10. X25. X50. X75. X90. acf
output(s): oxy.METS.calculated options were - skip-layer connections  linear
output units
}
\usage{
reg.nn
}
\description{
Uniaxial neural network for use in the original uniaxial Sojourn method
}
\keyword{datasets}
