% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{scal.1}
\alias{scal.1}
\title{Scaling coefficients for triaxial nnetinputs}
\format{
numeric vector of size 25
}
\usage{
scal.1
}
\description{
Scaling coefficients for triaxial nnetinputs
}
\keyword{datasets}
