% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{scal}
\alias{scal}
\title{Scaling coefficients for uniaxial nnetinputs}
\format{
numeric vector of size 6
}
\usage{
scal
}
\description{
Scaling coefficients for uniaxial nnetinputs
}
\keyword{datasets}
