% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sojourn_Data.R
\docType{package}
\name{Sojourn.Data}
\alias{Sojourn.Data}
\alias{Sojourn.Data-package}
\title{Sojourn.Data: Models for Sojourn Accelerometer Methods}
\description{
Sojourn methods rely on large objects, which take up too much space in an
ordinary package. Thus, the objects are stored in this data-only package,
meant to complement the Sojourn package.
}
