% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{youth_grids}
\alias{youth_grids}
\title{Data frame containing grid values for the youth Sojourn method}
\format{data frame with 4 rows and 14 columns}
\usage{
youth_grids
}
\description{
Data frame containing grid values for the youth Sojourn method
}
\keyword{datasets}
