% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{youth_wristCounts}
\alias{youth_wristCounts}
\title{Neural network for youth Sojourn method, taking activity count data from
non-dominant-wrist-worn monitors}
\format{
From print(youth_wristCounts):
\describe{
a 9-15-3 network with 198 weights inputs: Age SexM BMI VM_Q10 VM_Q25 VM_Q50
VM_Q75 VM_Q90 VM_lag1 output(s): .outcome options were - softmax modelling
decay=0.1
}
}
\usage{
youth_wristCounts
}
\description{
Neural network for youth Sojourn method, taking activity count data from
non-dominant-wrist-worn monitors
}
\keyword{datasets}
