% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{youth_wristRaw}
\alias{youth_wristRaw}
\title{Neural network for youth Sojourn method, taking raw accelerometer data from
non-dominant-wrist-worn monitors}
\format{
From print(youth_wristRaw):
\describe{
a 9-15-3 network with 198 weights inputs: Age SexM BMI ENMO_Q10 ENMO_Q25
ENMO_Q50 ENMO_Q75 ENMO_Q90 ENMO_lag1 output(s): .outcome options were -
softmax modelling  decay=0.1
}
}
\usage{
youth_wristRaw
}
\description{
Neural network for youth Sojourn method, taking raw accelerometer data from
non-dominant-wrist-worn monitors
}
\keyword{datasets}
